workaround twitter API changes, maybe
seems I'm now limited to 25 requests per 24 hours ( https://developer.twitter.com/en/portal/products/free ), and the API response for exceeding that is to return a Twitter::Error::Unauthorized? ¯\_(ツ)_/¯

so, reduce twitter refresh to hourly at most, and try to handle this error in the background job without emailing me about it.