initial work on getting playstation trophies through the playstation API
cannot currently get an auth token, as the https://ca.account.sony.com/api/authz/v3/oauth/token endpoint gives a 429 (too many requests) error no matter how many times I log out and generate a new npsso [despite similar code on github saying that this is the fix], or how long I wait between attempts at mucking with it. I even created a new API-only PSN account (because I thought 2FA might be part of the problem and I refuse to disable it) and got the same result. my current running theory is that something in the API has changed (as all of this is undocumented and Sony is actively antogonistic towards it) and others who already have auth and refresh tokens are not running into it yet, but I might also just be missing something very obvious.

references:
  https://github.com/andshrew/PlayStation-Trophies
  https://github.com/Tustin/psn-php
  https://github.com/isFakeAccount/psnawp