get trophies for most recent session in a PSN game
since the last played time reported by PSN is when a game session was last started, any trophies earned during that session wouldn't be looked up (after the first check).

so instead, look up trophies for all new games and all games where the reported last played time is within the last 24 hours (I'm working on the assumption that my days of >24 hour gaming sessions are in the past).