prevent overzealous last played time for nintendo platforms
in 48511fdb2f84654aeb48118b8bbd7c27a078f1b2 I was correct about the lastUpdatedAt being the last time the console "phoned home", but was incorrect about when the console would actually do that (I assumed just while it was in use, but it turns out it just does it periodically all the damn time). so instead I'll only update the started playing time on a RecentGame the first time that game is seen each day, which is probably as close to accurate as I'm likely to get.