prevent overzealous last played time for nintendo platforms
in ed7dc4232b4d7726f18994067b1f6b9448c521aa I was correct about the lastUpdatedAt being the last time the console "phoned home", but was incorrect about when the console would actually do that (I assumed just while it was in use, but it turns out it just does it periodically all the damn time). so instead I'll only update the started playing time on a RecentGame the first time that game is seen each day, which is probably as close to accurate as I'm likely to get.