ease comparison for determining which games are GOTY eligible for the current year
there really isn't a good reason to check that the creation time is after the start of the year, as that is already covered by the release date, and this comparison excluded a game (deltarune) which had enough additions this year to be a GOTY contender (chapters 3&4), but the first half was originally released in 2019, which is when the RecentGame was created. this sort of date/release nonsense is hopefully rare.